requirements.txt
aws-cdk.assets==1.2.0
aws-cdk.aws-apigateway==1.2.0
aws-cdk.aws-applicationautoscaling==1.2.0
aws-cdk.aws-autoscaling==1.2.0
aws-cdk.aws-autoscaling-common==1.2.0
aws-cdk.aws-autoscaling-hooktargets==1.2.0
aws-cdk.aws-certificatemanager==1.2.0
aws-cdk.aws-cloudformation==1.2.0
aws-cdk.aws-cloudfront==1.2.0
aws-cdk.aws-cloudwatch==1.2.0
aws-cdk.aws-cloudwatch-actions==1.2.0
aws-cdk.aws-codebuild==1.2.0
aws-cdk.aws-codecommit==1.2.0
aws-cdk.aws-codepipeline==1.2.0
aws-cdk.aws-ec2==1.2.0
aws-cdk.aws-ecr==1.2.0
aws-cdk.aws-ecr-assets==1.2.0
aws-cdk.aws-ecs==1.2.0
aws-cdk.aws-elasticloadbalancing==1.2.0
aws-cdk.aws-elasticloadbalancingv2==1.2.0
aws-cdk.aws-events==1.2.0
aws-cdk.aws-events-targets==1.2.0
aws-cdk.aws-iam==1.2.0
aws-cdk.aws-kms==1.2.0
aws-cdk.aws-lambda==1.2.0
aws-cdk.aws-logs==1.2.0
aws-cdk.aws-route53==1.2.0
aws-cdk.aws-route53-targets==1.2.0
aws-cdk.aws-s3==1.2.0
aws-cdk.aws-s3-assets==1.2.0
aws-cdk.aws-secretsmanager==1.2.0
aws-cdk.aws-servicediscovery==1.2.0
aws-cdk.aws-sns==1.2.0
aws-cdk.aws-sns-subscriptions==1.2.0
aws-cdk.aws-sqs==1.2.0
aws-cdk.aws-ssm==1.2.0
aws-cdk.aws-stepfunctions==1.2.0
aws-cdk.core==1.2.0
aws-cdk.custom-resources==1.2.0
aws-cdk.cx-api==1.2.0
aws-cdk.region-info==1.2.0
